add tests

+++ src/tasks/tasks.service.ts
@@ -1,3 +1,4 @@
+import { User } from './../auth/user.entity';
 import { TaskStatus } from './task-status.enum';
 import { TaskRepository } from './task.repository';
 import { Injectable, NotFoundException } from '@nestjs/common';
@@ -15,24 +16,12 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) { }
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return await this.taskRepository.getTasks(filterDto)
+    async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+        return await this.taskRepository.getTasks(filterDto, user)
     }
 
-    // getAllTasks(): Task[] {
-    //     return this.tasks
-    // }
-
-    // getTasksWithFilter(filterDto: GetTasksFilterDto): Task[] {
-    //     const { status, search } = filterDto;
-    //     let tasks = this.getAllTasks();
-
-
-    //     return tasks;
-    // }
-
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
         if (!found) {
             throw new NotFoundException(`Task with ID ${id} not found`);
@@ -40,20 +29,20 @@ export class TasksService {
 
         return found;
     }
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });
 
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID ${id} not found`);
         }
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return await this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return await this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
 
         return await task.save();
